requirements.txt
html-testRunner==1.2.1
locust==2.16.1
pip==22.3.1
PySocks==1.7.1
python-dotenv==1.0.0
selenium==4.12.0
wheel==0.38.4